+++ src/swap/defi/uni-v2-based/plugins/spookySwap.ts
@@ -6,7 +6,8 @@ import {
   EdgeSwapPlugin,
   EdgeSwapQuote,
   EdgeSwapRequest,
-  EdgeTransaction
+  EdgeTransaction,
+  SwapCurrencyError
 } from 'edge-core-js/types'
 import { ethers } from 'ethers'
 
@@ -22,7 +23,8 @@ import { getInOutTokenAddresses, InOutTokenAddresses } from '../../defiUtils'
 import {
   getFtmProvider,
   makeSpookySwapRouterContract,
-  makeWrappedFtmContract
+  makeWrappedFtmContract,
+  WFTM_TOKEN_ADDRESS
 } from '../uniV2Contracts'
 import { getSwapAmounts, getSwapTransactions } from '../uniV2Utils'
 
@@ -50,27 +52,22 @@ export function makeSpookySwapPlugin(
     request: EdgeSwapRequestPlugin,
     uid: string
   ): Promise<SwapOrder> => {
-    const {
-      fromWallet,
-      toWallet,
-      fromCurrencyCode,
-      toCurrencyCode,
-      quoteFor
-    } = request
+    const { fromWallet, toWallet, fromTokenId, toTokenId, quoteFor } = request
 
     // Sanity check: Both wallets should be of the same chain.
     if (
-      fromWallet.currencyInfo.currencyCode !==
-      toWallet.currencyInfo.currencyCode
+      fromWallet.currencyInfo.currencyCode !== 'fantom' ||
+      toWallet.currencyInfo.currencyCode !== 'fantom'
     )
-      throw new Error(`${swapInfo.displayName}: Mismatched wallet chain`)
+      throw new SwapCurrencyError(swapInfo, request)
 
     // Parse input/output token addresses. If either from or to swap sources
     // are for the native currency, convert the address to the wrapped equivalent.
     const inOutAddresses: InOutTokenAddresses = getInOutTokenAddresses(
-      fromWallet.currencyInfo,
-      fromCurrencyCode,
-      toCurrencyCode
+      fromWallet.currencyConfig,
+      WFTM_TOKEN_ADDRESS,
+      fromTokenId,
+      toTokenId
     )
     const { fromTokenAddress, toTokenAddress, isWrappingSwap } = inOutAddresses